project/backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1
requests==2.31.0
moviepy==1.0.3
pyttsx3==2.90
wikipedia==1.4.0
duckduckgo-search==4.4
beautifulsoup4==4.12.3
aiofiles==23.2.1  # Needed for serving static files with FastAPI
